requirements.txt
google_api_python_client==2.128.0
google_auth_oauthlib==1.2.0
langchain_community==0.0.38
langchain_groq==0.1.3
matplotlib==3.8.4
numpy==1.26.4
pandas==1.5.3
pandasai==2.0.37
plotly==5.22.0
protobuf==4.25.3
python-dotenv==1.0.1
streamlit==1.34.0
streamlit-aggrid==1.0.4.post3
streamlit_authenticator==0.3.2
streamlit_option_menu==0.3.12
streamlit_extras==0.4.2
xlrd==2.0.1
PyYAML==6.0.1
gspread==6.1.0
gspread_dataframe==3.3.1
